session: return explicit nil to avoid non-nil interface value
A interface has two elements, a type and a value. A interface value is
only nil if both elements are nil. However if we return AccepStream()
directly, net.Conn will be: (net.Conn, nil). Even if the inner value is
nil, this value will be non-nil. So when Open() and Accept() return an
error, the value (in our example net.Conn) will be always non-nil.

This commit fixes it by explicit returning a nil. More detail about
this can be found: http://golang.org/doc/faq#nil_error