Fix failing partial read on last frame of closed session
If the sending side of a stream closes both their stream and their
session, a partial read of the final frame on the receiving side may
fail because the receiving side tries to send a window update on a
closed session.

This commit changes Stream.Read() so that it ignores "session
shutdown" errors return by Stream.sendWindowUpdate(). This will let
Read() calls consume all of the final frame.