Implement timeouts for OpenStream
Currently streams can be in the streamSYNSent state for an unbounded
amount of time when the remote is unable to reply and has not closed the
connection via FIN/RST.

This commit adds a deadline to receiving an ACK after a SYN is sent on
OpenStream.

The timeout is configurable per-session with StreamOpenTimeout, and logs
an error if it is reached.